Finish split of datasources and vis_flags_weights
